[py_trees] nosetest fixes, sequence abort bugfix, behaviour update return guard and a new meta.
This fixes several problems. The nosetests were out of date, so this fixes them. It also
caught a bug in the sequence where it would set everything to invalid whenever it triggered the
abort function.

Adds a new guard to behaviour updates, catching when a users behaviour does not set a result,
or provides something that is not in enum Status.

And a new meta behaviour...failure is success, when you need to march on through a sequence
regardless.