[rqt_behaviour_tree] new package for viewing behaviour trees
Currently has an undiagnosed segfault issue. Occurs whenever the tree is
updated, and sometimes arbitrarily. It happens sometime in the redraw_graph_view
function. Sometimes it's a segfault, sometimes multiple free. Seems to be
something to do with _scene.clear(). Removing that and replacing it with the
initialisation of the scene from the init function means that things aren't
displayed properly any more, but it no longer crashes. There is some mention of
the QTimer only being able to be run in the QThread, and the only place that is
used is in rqt_gui.main, as far as I can tell. Must be something to do with
elements being drawn again when they've been deleted, but can't get in there to
debug.