[meta] bugfix timeout decorators, fixes #38
They stopped triggering initialise() after #34. The key point here
is to manually trigger the initialise externally and let the underlying
initialise be triggered by the underlying behaviour. This needed
the alias to initialise in the imposter to be removed. It should
not be there after #34, but keep an eye that this does not break anything
else.