Isolate test and doc dependencies from installed dependencies (#72)
Makes use of the `extra_requires` variable in `setup.py`. It also updates the virtualenv bash scripts to make use of `setup.py` for installing dependencies instead of the ugly duckling - rtd-requirements.txt which we want to go away if at some point read-the-docs gives us a way to use the one source of truth for dependencies - setup.py.