[behaviour] setup now accepts **kwargs
This allows setup of runtime components that must be instantiated
from outside the behaviour, but can't be passed in at init time.
Examples: a simulator instance, or a ros2 node to share to the tree.

Not an ideal solution, given the key conflicts that could occur.
The alternatives are less appealing.