[composites] pulled up short on dropping the stop in tick here, just comments
This is very confusing when children get terminated before they even
get a chance to initialise. Ostensibly, the parallel still needs
to reset the status of the children and it can probably do just
that, but not pulling the plug on it in this update just in case ...
there are consequences.