minor fixes to 1.3.x (#240)
* [blackboard] no copying/pickling, just use handles

Copying, deepcopying often result in pickling happening, which will
almost always get you into trouble if you are storing objects on the
blackboard that are python bound (i.e. c/c++ under the hood).

Just use handles for now, it's eminently usable and tree ticks
as we have designed for here dont have to handle the concurrency issues.

* [composites] pulled up short on dropping the stop in tick here, just
comments

This is very confusing when children get terminated before they even
get a chance to initialise. Ostensibly, the parallel still needs
to reset the status of the children and it can probably do just
that, but not pulling the plug on it in this update just in case ...
there are consequences.