[composites] use typing.Sequence for children argument
The Python type system considers `typing.List` to be mutable, and thus
invariant on its child type. This doesn't let us pass `Behaviour`
subclasses to composites through the `children` parameter, forcing the
use of `composite.add_child(child)` or the use of `typing.cast`. Since
the `children` list is never mutated, it's better to use
`typing.Sequence` here which is covariant and therefore typechecks when
a list of `Behaviour` subclasses are passed in.